+++ CHANGELOG.md
@@ -4,13 +4,14 @@
 - [FEATURE] When i18n installed but the validation translation not defined use
   the default one (#37)
 - [FEATURE] Add ability to provide custom error and success classes for validation, thanks @kcin1018!
+- [FEATURE] Make it so that you can override `form-control` class on inputs (#53)
 
 # 2.1.0
 - [INTERNAL] Update to ember 2.11
 - [FEATURE] Add ability to create custom inputs, thanks @bardzusny!
 - [FEATURE] Make it possible to add classes directly to the input component
   (#40)
-  
+
 # 2.0.1
 
 - [INTERNAL] Update ember-cli-release and ember-one-way-controls to remove deprecations, thanks @maxwerr